refactor: 도서 목록 페이지의 PageLayout을 하위 컴포넌트로 분리
