+++ src/app/books/[list]/layout.tsx
@@ -1,11 +1,6 @@
-import { H1 } from '@/components';
 import { BOOKS_TITLE } from '@/constants';
 
-import styles from './layout.module.scss';
-
-export const TITLE: Record<string, string> = {
-  popular: '인기 대출 도서',
-};
+import { BookListPageLayout } from './_components/PageLayout';
 
 export type BooksPageParams = Promise<{
   list: string;
@@ -17,20 +12,6 @@ export const generateMetadata = async ({ params }: { params: BooksPageParams }) 
   return {
     title: `BOOKER - ${BOOKS_TITLE[list]}`,
   };
-};
-
-interface BookListPageLayoutProps {
-  list: string;
-  children: React.ReactNode;
-}
-
-export const BookListPageLayout = ({ list, children }: BookListPageLayoutProps) => {
-  return (
-    <div className={styles.layout}>
-      <H1 contents={TITLE[list]} />
-      {children}
-    </div>
-  );
 };
 
 interface LayoutProps {
